refactor: Renamed interface
